+++ NorthWest.py
